Revert "Task 6385486: Restructured Adapter infra into a inheritance model and defined a new adapter for RS Backup Mgmt Client"
This reverts commit 984d00b06d9a5c12e859452638797c467fd13488.
