+++ src/ResourceManager/AzureBackup/Commands.AzureBackup/AzureBackupClientAdapter/AzureBackupClientAdapter.cs
@@ -21,8 +21,23 @@ using System.Threading;
 
 namespace Microsoft.Azure.Commands.AzureBackup.ClientAdapter
 {
-    public partial class AzureBackupClientAdapter : ClientAdapterBase
+    public partial class AzureBackupClientAdapter
     {
+        /// <summary>
+        /// Cloud credentials for client calls
+        /// </summary>
+        private SubscriptionCloudCredentials cloudCreds { get; set; }
+
+        /// <summary>
+        /// Base URI for client calls
+        /// </summary>
+        private Uri baseURI { get; set; }
+
+        /// <summary>
+        /// Client request id.
+        /// </summary>
+        private string clientRequestId;
+
         /// <summary>
         /// Azure vault backup client to talk to IdMgmt.
         /// </summary>
@@ -34,6 +49,12 @@ namespace Microsoft.Azure.Commands.AzureBackup.ClientAdapter
         private BackupServicesManagementClient azureBackupClient;
 
         /// <summary>
+        /// Cancellation Token Source
+        /// </summary>
+        private CancellationTokenSource cancellationTokenSource = new CancellationTokenSource();
+        private CancellationToken CmdletCancellationToken;
+
+        /// <summary>
         /// Get Azure backup client.
         /// </summary>
         private BackupVaultServicesManagementClient AzureBackupVaultClient
@@ -42,7 +63,7 @@ namespace Microsoft.Azure.Commands.AzureBackup.ClientAdapter
             {
                 if (this.azureBackupVaultClient == null)
                 {
-                    this.azureBackupVaultClient = AzureSession.ClientFactory.CreateCustomClient<BackupVaultServicesManagementClient>(CloudCreds, BaseURI);
+                    this.azureBackupVaultClient = AzureSession.ClientFactory.CreateCustomClient<BackupVaultServicesManagementClient>(cloudCreds, baseURI);
                 }
 
                 return this.azureBackupVaultClient;
@@ -58,7 +79,7 @@ namespace Microsoft.Azure.Commands.AzureBackup.ClientAdapter
             {
                 if (this.azureBackupClient == null)
                 {
-                    this.azureBackupClient = AzureSession.ClientFactory.CreateCustomClient<BackupServicesManagementClient>(CloudCreds, BaseURI);
+                    this.azureBackupClient = AzureSession.ClientFactory.CreateCustomClient<BackupServicesManagementClient>(cloudCreds, baseURI);
                 }
 
                 return this.azureBackupClient;
@@ -66,14 +87,32 @@ namespace Microsoft.Azure.Commands.AzureBackup.ClientAdapter
         }
 
         public AzureBackupClientAdapter(SubscriptionCloudCredentials creds, Uri baseUri)
-            : base(creds, baseUri) { }
+        {
+            cloudCreds = creds;
+            baseURI = baseUri;
+
+            RefreshClientRequestId();
+
+            // Temp code to be able to test internal env.
+            ServicePointManager.ServerCertificateValidationCallback = delegate { return true; };
+        }
+
+        public void RefreshClientRequestId()
+        {
+            clientRequestId = Guid.NewGuid().ToString() + "-" + DateTime.Now.ToUniversalTime().ToString("yyyy-MM-dd HH:mm:ssZ") + "-PS";
+        }
+
+        public string GetClientRequestId()
+        {
+            return clientRequestId;
+        }
 
         internal CustomRequestHeaders GetCustomRequestHeaders()
         {
             var hdrs = new CustomRequestHeaders()
             {
                 // ClientRequestId is a unique ID for every request to backend service.
-                ClientRequestId = this.ClientRequestId,
+                ClientRequestId = this.clientRequestId,
             };
 
             return hdrs;
