--- src/ResourceManager/AzureBackup/Commands.AzureBackup/AzureBackupClientAdapter/RecoveryServicesAdapters/ContainerAdapter.cs
@@ -1,56 +1,0 @@
-﻿// ----------------------------------------------------------------------------------
-//
-// Copyright Microsoft Corporation
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// http://www.apache.org/licenses/LICENSE-2.0
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-// ----------------------------------------------------------------------------------
-
-using System;
-using System.Management.Automation;
-using System.Collections.Generic;
-using System.Xml;
-using Microsoft.WindowsAzure.Commands.Utilities.Common;
-using Microsoft.Azure.Common.Authentication;
-using Microsoft.Azure.Common.Authentication.Models;
-using System.Threading;
-using Hyak.Common;
-using Microsoft.Azure.Commands.AzureBackup.Properties;
-using System.Net;
-using System.Linq;
-using Microsoft.WindowsAzure.Management.Scheduler;
-using Microsoft.Azure.Commands.AzureBackup.Models;
-using Microsoft.Azure.Management.RecoveryServices.Backup.Models;
-
-namespace Microsoft.Azure.Commands.AzureBackup.ClientAdapter
-{
-    public partial class AzureRsClientAdapter
-    {
-        /// <summary>
-        /// Fetches protection containers in the vault according to the query params
-        /// </summary>
-        /// <param name="parameters"></param>
-        /// <returns></returns>
-        public IEnumerable<ProtectionContainerResource> ListContainers(string resourceGroupName, string resourceName, ProtectionContainerListQueryParams queryParams)
-        {
-            var listResponse = AzureRSClient.Container.ListAsync(resourceGroupName, resourceName, queryParams, GetCustomRequestHeaders(), CmdletCancellationToken).Result;
-            return listResponse.ItemList.ProtectionContainers;
-        }
-
-        /// <summary>
-        /// Triggers refresh of container catalog in service
-        /// </summary>
-        /// <returns></returns>
-        public BaseRecoveryServicesJobResponse RefreshContainers(string resourceGroupName, string resourceName)
-        {
-            var response = AzureRSClient.Container.RefreshAsync(resourceGroupName, resourceName, GetCustomRequestHeaders(), AzureFabricName, CmdletCancellationToken).Result;
-            return response;
-        }
-    }
-}
